\name{l2010}
\alias{l2010}
\title{World lightning strike data from 2010}
\description{
 A list of two 50x100 matrices containing most of the world lightning strike 
 data from 2010. It was produced by \samp{makeDensityMatrix} from 171 file 
 (about 3 Gb) of data consisting in two geographic coordinates for the 
 approximate location of each recorded strike and an estimated intensity of the 
 strike in kVA.
}
\usage{
 data(l2010)
}
\keyword{misc}
