\name{oksoil}
\alias{oksoil}
\title{Soil texture data from Oklahoma soils}
\description{
 A set of 125 soil texture measurements from Oklahoma soils.
}
\usage{
 oksoil
}
\source{
 T.H. Skaggs, L.M. Arya, P.J. Shouse and B.P. Mohanty (2001)
 Estimating Particle-Size Distribution from Limited Soil Texture Data.
 Soil Science Society of America Journal 65:1038-1044.
}
\keyword{misc}
