\name{piper.coords}
\alias{piper.coords}
\title{Calculate point coordinates for a Piper diagram.}
\description{Calculate point coordinates for a Piper diagram.}
\usage{
 piper.coords(ions,plot.sep=0.15)
}
\arguments{
 \item{ions}{data frame containing ion concentrations.}
 \item{plot.sep}{The offsets for the ions.}
}
\value{A data frame containing the coordinates of the points.}
\author{Mike Cheetham}
\seealso{\samp{\link{piper.diagram}}}
\keyword{misc}
