\name{piper.set}
\alias{piper.set}
\title{Calculate ion offsets.}
\description{Calculates the offsets for anions and cations in the Piper
 diagram.}
\usage{
 piper.set(plot.sep=0.15)
}
\arguments{
 \item{plot.sep}{The offsets of the ions in user units.}
}
\details{  
 \samp{piper.set} calculates the ion offsets in the Piper diagram.
}
\value{The offsets.}
\author{Mike Cheetham}
\seealso{\samp{\link{piper.diagram}}}
\keyword{misc}
