\name{plotrix-package}
\alias{plotrix-package}
\alias{plotrix}
\docType{package}
\title{
Specialized plots and plotting accessories
}
\description{
A large number of specialized plots and accessory functions like color
scaling, text placement and legends.
}
\details{
%% \tabular{ll}{
%% Package: \tab plotrix\cr
%% Version: \tab 3.8-1\cr
%% Date: \tab 2021-09-08\cr
%% License: \tab GPL (>=2)\cr
%% Packaged: \tab 2021-09-08 03:45:00 UTC; root\cr
%% Built: \tab R 4.0.3; ; 2021-09-08 03:45:00 UTC; linux\cr
%% }
 The plotrix package is intended to provide a method for getting many sorts of
 specialized plots quickly, yet allow easy customization of those plots
 without learning a great deal of specialized syntax. There are three major
 aims that can be represented as follows:

Fast foods

Think of plotrix as a graphics vending machine or fast graphics cafe. You walk
 in, make your choice and get your lunch. It may not be exactly the lunch you
 want, but you do get a pretty good lunch, fast. You can get junk food or
 health food, you make the choice.

Hot rods

You can customize plotrix as much as you want. Like the ageing machinery that
 is usually bolted into hot rods, the base graphics package is fairly easy to
 understand. plotrix is modular. You can create a frame for your plot, then
 you can add whatever bits you like to it instead of just taking the default
 plot that is available. You can have wide wheels and chromed exhaust pipes if
 you want.

No black boxes

If you want to go from pushing the fast food button to hot rodding, it's not
 hard. The source code in the functions is written to be understood. If
 something goes wrong, you can usually find where it happened right away and
 work on it. This means that you can learn about how the functions do what
 they do rather than just what they do. So that's how to write recursive
 functions in R!

Because plotrix encourages users to learn how it works, you usually begin to
 do so pretty quickly. Users often decide to write their own versions of
 plotrix functions and sometimes they contribute the results back into plotrix.
 You may find that you like other graphics systems like grid or lattice better.
 That's great, because one idea behind plotrix is that if you get into R and
 can get things done quickly and easily, you'll stick with it and soon want to
 get things done your way. 
}
\author{
Jim Lemon <drjimlemon@gmail.com>, and many others

Maintainer: Jim Lemon <drjimlemon@gmail.com>
}
\keyword{ package }
