\name{soils}
\alias{soils}
\title{Soil texture data from 125 soils}
\description{
 A set of 125 soil texture measurements from soils from various parts of the
  world.
}
\usage{
 data(soils)
}
\source{
 T.H. Skaggs, L.M. Arya, P.J. Shouse and B.P. Mohanty (2001)
 Estimating Particle-Size Distribution from Limited Soil Texture Data.
 Soil Science Society of America Journal 65:1038-1044.
}
\keyword{misc}
